10.1 move Votes activity from Answer to concern WithVotes
